requirements.txt
streamlit
pandas
openpyxl
plotly
yfinance
google-generativeai
duckduckgo-search
streamlit-chat
